+++ HAL/src/hal_gpio.c
@@ -55,7 +55,7 @@ Status_t HAL_GPIO_Init(void) {
  * @param level The desired output level (0 for low, 1 for high).
  * @return E_OK on success, or an error code.
  */
-Status_t HAL_GPIO_SetLevel(gpio_num_t gpio_num, uint32_t level) {
+Status_t HAL_GPIO_SetLevel(gpio_num_t gpio_num, uint8_t level) {
     esp_err_t ret = gpio_set_level(gpio_num, level);
     if (ret != ESP_OK) {
         ESP_LOGE(TAG, "Failed to set GPIO %d level to %d: %s", gpio_num, level, esp_err_to_name(ret));
@@ -70,7 +70,7 @@ Status_t HAL_GPIO_SetLevel(gpio_num_t gpio_num, uint32_t level) {
  * @param level_out Pointer to store the read level (0 for low, 1 for high).
  * @return E_OK on success, or an error code.
  */
-Status_t HAL_GPIO_GetLevel(gpio_num_t gpio_num, uint32_t *level_out) {
+Status_t HAL_GPIO_GetLevel(gpio_num_t gpio_num, uint8_t *level_out) {
     if (level_out == NULL) {
         ESP_LOGE(TAG, "HAL_GPIO_GetLevel: level_out pointer is NULL for GPIO %d.", gpio_num);
         return E_INVALID_PARAM;
